Testes adicionados e ajuste na formatação do size

+++ src/main/kotlin/com/danielxavier/invoiceEntry/application/usecase/dto/ObjectResponse.kt
@@ -3,5 +3,5 @@ package com.danielxavier.invoiceEntry.application.usecase.dto
 data class ObjectResponse(
     val key: String,
     val type: String?,
-    val size: Long
+    val size: Double
 )
